pyproject.toml
[project]
name = "litgpt-project"
version = "0.1.0"
description = "MoE TinyStories with LitGPT - Mixture of Experts Language Model"
requires-python = ">=3.12,<3.14"
dependencies = [
    "torch>=2.10.0,<2.11",
    "torchvision>=0.25.0,<0.26",
    "torchaudio>=2.10.0,<2.11",
    "litgpt",
    "lightning",
    "pyyaml",
    "requests",
    "wandb",
    "tokenizers",
    "datasets",
    "accelerate",
    "transformers",
    "safetensors",
    "sentencepiece",
    "bitsandbytes",
    "litdata>=0.2.59",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10",
    "pytest-cov>=4.0,<6",
    "ruff>=0.6,<1",
]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 127
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"run_train.py" = ["E402", "I001"]
